Module ggez::conf

Functions for loading and saving engine configurations.

A Conf struct is used to specify hardware setup stuff used to create the window and other context information.

By default, a ggez game will search it's resource paths for a /conf.toml file, and load values from it on Context creation; this file doesn't have to be complete (you can fill in some fields and have the rest be default).

Conversely, when writing Conf to a TOML file (via Conf::to_toml_file()), fields with default values will not appear in the output.

Default values can be found in documentation of structs of the module.

Here's an example config and it's equivalent TOML:

extern crate ggez;
extern crate toml;
use ggez::conf::*;

assert_eq!(
    Conf {
        window_mode: WindowMode::default()
            .always_on_top(true)
            .dimensions(640, 480)
            .min_dimensions(640, 480)
            .fullscreen_type(FullscreenType::Desktop(MonitorId::Index(1))),
        window_setup: WindowSetup::default()
            .title("Testing, testing, one, two")
            .samples(NumSamples::Four)
            .vsync(false),
        backend: Backend::OpenGL(4, 6),
    },
    toml::from_str::<Conf>(
        r#"
        [window_mode]
        always_on_top = true

        [window_mode.dimensions]
        width = 640
        height = 480

        [window_mode.min_dimensions]
        width = 640
        height = 480

        [window_mode.fullscreen]
        type = "Desktop"
        monitor = 1

        [window_setup]
        title = "Testing, testing, one, two"
        samples = "Four"
        vsync = false

        [backend]
        type = "OpenGL"
        version_major = 4
        version_minor = 6
        "#
    ).unwrap(),
);

Structs

Conf
Structure containing configuration data for the engine.

WindowMode
A builder structure containing window settings that can be set at runtime and changed with graphics::set_mode().

WindowSetup
A builder structure containing window settings that must be set at init time and (mostly) cannot be changed afterwards.

Enums

Backend
Possible graphics backends. Currently, only OpenGL itself is supported, but this lets you specify the version numbers.

FullscreenType
Possible fullscreen modes.

MonitorId
Identifies a monitor connected to the system.

NumSamples
Possible number of samples for multisample anti-aliasing (MSAA).