Struct ggez::conf::Conf

pub struct Conf {
    pub window_mode: WindowMode,
    pub window_setup: WindowSetup,
    pub backend: Backend,
}

Structure containing configuration data for the engine.

Defaults:

extern crate ggez;
use ggez::conf::{Backend, Conf, WindowMode, WindowSetup};

assert_eq!(
    Conf::default(),
    Conf {
        window_mode: WindowMode::default(),
        window_setup: WindowSetup::default(),
        backend: Backend::OpenGL(3, 2),
    },
);

Fields
window_mode: WindowMode

Window setting information that can be set at runtime.

window_setup: WindowSetup

Window setting information that must be set at init-time.

backend: Backend

Graphics backend.

Methods

impl Conf

pub fn new() -> Self

Creates a new default config. Identical to Conf::default().

pub fn from_toml_file<R: Read>(file: &mut R) -> GameResult<Conf>

Load a TOML file from the given Read and attempts to parse a Conf from it.

pub fn to_toml_file<W: Write>(&self, file: &mut W) -> GameResult

Saves the Conf to the given Write object, formatted as valid TOML.