Struct ggez::conf::WindowMode

pub struct WindowMode {
    pub dimensions: (u32, u32),
    pub min_dimensions: Option<(u32, u32)>,
    pub max_dimensions: Option<(u32, u32)>,
    pub fullscreen_type: FullscreenType,
    pub maximized: bool,
    pub hidden: bool,
    pub borderless: bool,
    pub always_on_top: bool,
}

A builder structure containing window settings that can be set at runtime and changed with graphics::set_mode().

Defaults:

extern crate ggez;
use ggez::conf::{FullscreenType, WindowMode};

assert_eq!(
    WindowMode::default(),
    WindowMode {
        dimensions: (800, 600),
        min_dimensions: None,
        max_dimensions: None,
        fullscreen_type: FullscreenType::Off,
        maximized: false,
        hidden: false,
        borderless: false,
        always_on_top: false,
    },
);

Fields
dimensions: (u32, u32)

Window's inner dimensions (drawable area).

min_dimensions: Option<(u32, u32)>

Window's minimum dimensions; None is no limit.

max_dimensions: Option<(u32, u32)>

Window's maximum dimensions; None is no limit.

fullscreen_type: FullscreenType

Fullscreen type.

maximized: bool

Whether or not to maximize the window.

hidden: bool

Whether or not the window is hidden.

borderless: bool

Whether or not to show window decorations.

always_on_top: bool

Whether or not the window is pinned to always be on top of other windows.

Methods

impl WindowMode

pub fn dimensions(self, width: u32, height: u32) -> Self

Set default window size, or screen resolution in true fullscreen mode.

pub fn min_dimensions(self, width: u32, height: u32) -> Self

Set minimum window dimensions for windowed mode.

pub fn max_dimensions(self, width: u32, height: u32) -> Self

Set maximum window dimensions for windowed mode.

pub fn fullscreen_type(self, fullscreen_type: FullscreenType) -> Self

Set the fullscreen type.

pub fn maximized(self, maximized: bool) -> Self

Set if window should be maximized.

pub fn hidden(self, hidden: bool) -> Self

Set if window should be hidden.

pub fn borderless(self, borderless: bool) -> Self

Set whether or not to show window decorations.

pub fn always_on_top(self, always_on_top: bool) -> Self

Set if window should pinned to always be on top of other windows.