Enum ggez::error::GameError

pub enum GameError {
    FilesystemError(String),
    ConfigError(String),
    EventLoopError(String),
    ResourceLoadError(String),
    ResourceNotFound(String, Vec<(PathBuf, GameError)>),
    RenderError(String),
    AudioError(String),
    WindowError(CreationError),
    IOError(Error),
    FontError(String),
    VideoError(String),
    ShaderProgramError(ProgramError),
    GamepadError(String),
    UnknownError(String),
}

An enum containing all kinds of game framework errors.

Variants
FilesystemError(String)

An error in the filesystem layout

ConfigError(String)

An error in the config file

EventLoopError(String)

Happens when an EventsLoopProxy attempts to wake up an EventsLoop that no longer exists.

ResourceLoadError(String)

An error trying to parse a resource

ResourceNotFound(String, Vec<(PathBuf, GameError)>)

Unable to find a resource; the Vec is the paths it searched for and associated errors

RenderError(String)

Something went wrong in the renderer

AudioError(String)

Something went wrong in the audio playback

WindowError(CreationError)

Something went wrong trying to create a window

IOError(Error)

Something went wrong trying to read from a file

FontError(String)

Something went wrong trying to load/render a font

VideoError(String)

Something went wrong applying video settings.

ShaderProgramError(ProgramError)

Something went wrong compiling shaders

GamepadError(String)

Something went wrong with Gilrs

UnknownError(String)

Something else happened; this is generally a bug.