Struct ggez::graphics::DrawParam

pub struct DrawParam { /* fields omitted */ }

A struct containing all the necessary info for drawing a Drawable.

This struct implements the Default trait, so to set only some parameter you can just do:

graphics::draw_ex(ctx, drawable, DrawParam{ dest: my_dest, .. Default::default()} )

Methods

impl DrawParam

pub fn new() -> Self

Create a new DrawParam with default values.

pub fn src(self, src: Rect) -> Self

Set the source rect

pub fn dest<P>(self, dest: P) -> Self where
P: Into<Point2<f32>>,

Set the dest point

pub fn color<C>(self, color: C) -> Self where
C: Into<Color>,

TODO

pub fn rotation(self, rotation: f32) -> Self

TODO

pub fn scale<V>(self, scale: V) -> Self where
V: Into<Vector2<f32>>,

TODO

pub fn offset<P>(self, offset: P) -> Self where
P: Into<Point2<f32>>,

TODO

pub fn shear<P>(self, shear: P) -> Self where
P: Into<Point2<f32>>,

TODO