Struct ggez::graphics::FillOptions

pub struct FillOptions {
    pub tolerance: f32,
    pub fill_rule: FillRule,
    pub compute_normals: bool,
    pub assume_no_intersections: bool,
    pub on_error: OnError,
    // some fields omitted
}

Parameters for the fill tessellator.

Fields
tolerance: f32

Maximum allowed distance to the path when building an approximation.

See Flattening and tolerance.

Default value: FillOptions::DEFAULT_TOLERANCE.

fill_rule: FillRule

Set the fill rule.

See the SVG specification. Currently, only the EvenOdd rule is implemented.

Default value: EvenOdd.

compute_normals: bool

Whether or not to compute the normal vector at each vertex.

When set to false, all generated vertex normals are equal to vector(0.0, 0.0). Not computing vertex normals can speed up tessellation and enable generating less vertices at intersections.

Default value: true.

assume_no_intersections: bool

A fast path to avoid some expensive operations if the path is known to not have any self-intersections.

Do not set this to true if the path may have intersecting edges else the tessellator may panic or produce incorrect results. In doubt, do not change the default value.

Default value: false.

on_error: OnError

What to do if the tessellator detects an error.

Methods

impl FillOptions

pub fn default() -> FillOptions

pub fn even_odd() -> FillOptions

pub fn tolerance(tolerance: f32) -> FillOptions

pub fn non_zero() -> FillOptions

pub fn with_tolerance(self, tolerance: f32) -> FillOptions

pub fn with_normals(self, normals: bool) -> FillOptions

pub fn assume_no_intersections(self) -> FillOptions

pub fn on_error(self, policy: OnError) -> FillOptions