Struct ggez::graphics::StrokeOptions

pub struct StrokeOptions {
    pub start_cap: LineCap,
    pub end_cap: LineCap,
    pub line_join: LineJoin,
    pub line_width: f32,
    pub miter_limit: f32,
    pub tolerance: f32,
    pub apply_line_width: bool,
    // some fields omitted
}

Parameters for the tessellator.

Fields
start_cap: LineCap

What cap to use at the start of each sub-path.

Default value: LineCap::Butt.

end_cap: LineCap

What cap to use at the end of each sub-path.

Default value: LineCap::Butt.

line_join: LineJoin

See the SVG specification.

Default value: LineJoin::Miter.

line_width: f32

Line width

Default value: StrokeOptions::DEFAULT_LINE_WIDTH.

miter_limit: f32

See the SVG specification.

Must be greater than or equal to 1.0. Default value: StrokeOptions::DEFAULT_MITER_LIMIT.

tolerance: f32

Maximum allowed distance to the path when building an approximation.

See Flattening and tolerance. Default value: StrokeOptions::DEFAULT_TOLERANCE.

apply_line_width: bool

Apply line width

When set to false, the generated vertices will all be positioned in the centre of the line. The width can be applied later on (eg in a vertex shader) by adding the vertex normal multiplied by the line with to each vertex position.

Default value: true.

Methods

impl StrokeOptions

pub fn default() -> StrokeOptions

pub fn tolerance(tolerance: f32) -> StrokeOptions

pub fn with_tolerance(self, tolerance: f32) -> StrokeOptions

pub fn with_line_cap(self, cap: LineCap) -> StrokeOptions

pub fn with_start_cap(self, cap: LineCap) -> StrokeOptions

pub fn with_end_cap(self, cap: LineCap) -> StrokeOptions

pub fn with_line_join(self, join: LineJoin) -> StrokeOptions

pub fn with_line_width(self, width: f32) -> StrokeOptions

pub fn with_miter_limit(self, limit: f32) -> StrokeOptions

pub fn dont_apply_line_width(self) -> StrokeOptions