Struct ggez::graphics::Text

pub struct Text { /* fields omitted */ }

Drawable text created from a Font.

Methods

impl Text

pub fn new(context: &mut Context, text: &str, font: &Font) -> GameResult<Text>

Renders a new Text from the given Font.

Note that this is relatively computationally expensive; if you want to draw text every frame you probably want to save it and only update it when the text changes.

pub fn width(&self) -> u32

Returns the width of the rendered text, in pixels.

pub fn height(&self) -> u32

Returns the height of the rendered text, in pixels.

pub fn contents(&self) -> &str

Returns the string that the text represents.

pub fn get_dimensions(&self) -> Rect

Returns the dimensions of the rendered text.

pub fn get_filter(&self) -> FilterMode

Get the filter mode for the the rendered text.

pub fn set_filter(&mut self, mode: FilterMode)

Set the filter mode for the the rendered text.

pub fn get_image(&self) -> &Image

Returns a reference to the Image contained by the Text object.

pub fn get_image_mut(&mut self) -> &mut Image

Returns a mutable reference to the Image contained by the Text object.

pub fn into_inner(self) -> Image

Unwraps the Image contained by the Text object.