Struct ggez::graphics::TextCached

pub struct TextCached { /* fields omitted */ }

Drawable text. Can be either monolithic, or consist of differently-formatted fragments.

Methods

impl TextCached

pub fn new<F>(fragment: F) -> GameResult<TextCached> where
F: Into<TextFragment>,

Creates a TextCached from a TextFragment.

pub fn new_empty() -> GameResult<TextCached>

Creates an empty TextCached.

pub fn add_fragment<F>(&mut self, fragment: F) -> &mut TextCached where
F: Into<TextFragment>,

Appends a TextFragment.

Important traits for &'a [u8]
impl<'a> Read for &'a [u8]impl<'a> Write for &'a mut [u8]

pub fn fragments(&self) -> &[TextFragment]

Returns a slice with all fragments, for reading.

Important traits for &'a [u8]
impl<'a> Read for &'a [u8]impl<'a> Write for &'a mut [u8]

pub fn fragments_mut(&mut self) -> &mut [TextFragment]

Returns a mutable slice with all fragments.

pub fn set_bounds<P>(&mut self, bounds: P, alignment: Align) -> &mut TextCached where
P: Into<Point2<f32>>,

Specifies rectangular dimensions to try and fit contents inside of, by wrapping, and alignment within the bounds.

pub fn set_font<FI>(
&mut self,
font_id: FI,
font_scale: Scale
) -> &mut TextCached where
FI: Into<FontId>,

Specifies text's font and font scale; used for fragments that don't have their own.

pub fn contents(&self) -> String

Returns the string that the text represents.

pub fn width(&self, context: &Context) -> u32

Returns the width of formatted and wrapped text, in screen coordinates.

pub fn height(&self, context: &Context) -> u32

Returns the height of formatted and wrapped text, in screen coordinates.

pub fn queue<P>(
&self,
context: &mut Context,
relative_dest: P,
color: Option<Color>
) where
P: Into<Point2<f32>>,

Queues the TextCached to be drawn by draw_queued(). relative_dest is relative to the DrawParam::dest passed to draw_queued(). Note, any TextCached drawn via graphics::draw() will also draw the queue.

pub fn queue_raw<'a, S, G>(
context: &mut Context,
section: S,
custom_layout: Option<&G>
) where
S: Into<Cow<'a, VariedSection<'a>>>,
G: GlyphPositioner,

Exposes gfx_glyph's GlyphBrush::queue() and GlyphBrush::queue_custom_layout(), in case ggez' API is insufficient.

pub fn draw_queued<D>(context: &mut Context, param: D) -> GameResult where
D: Into<PrimitiveDrawParam>,

Draws all of queue()d TextCached. DrawParam apply to everything in the queue; offset is in screen coordinates; color is ignored - specify it when queue()ing instead.