Trait ggez::graphics::BackendSpec

pub trait BackendSpec: Debug {
    type SurfaceType: Formatted;
    type Resources: Resources;
    type Factory: Factory<Self::Resources>;
    type CommandBuffer: CommandBuffer<Self::Resources>;
    type Device: Device<Resources = Self::Resources, CommandBuffer = Self::CommandBuffer>;
    fn raw_to_typed_shader_resource(
        texture_view: RawShaderResourceView<Self::Resources>
    ) -> ShaderResourceView<Self::Resources, <Self::SurfaceType as Formatted>::View> { ... }
}

A marker trait saying that something is a label for a particular backend, with associated gfx-rs types for that backend.

Associated Types

type SurfaceType: Formatted

Surface type

type Resources: Resources

gfx resource type

type Factory: Factory<Self::Resources>

gfx factory type

type CommandBuffer: CommandBuffer<Self::Resources>

gfx command buffer type

type Device: Device<Resources = Self::Resources, CommandBuffer = Self::CommandBuffer>

gfx device type

Provided Methods

fn raw_to_typed_shader_resource(
texture_view: RawShaderResourceView<Self::Resources>
) -> ShaderResourceView<Self::Resources, <Self::SurfaceType as Formatted>::View>

A helper function to take a RawShaderResourceView and turn it into a typed one based on the surface type defined in a BackendSpec