Trait ggez::graphics::Drawable

pub trait Drawable {
    fn draw_primitive(
        &self, 
        ctx: &mut Context, 
        param: PrimitiveDrawParam
    ) -> GameResult;

    fn set_blend_mode(&mut self, mode: Option<BlendMode>);

    fn get_blend_mode(&self) -> Option<BlendMode>;

    fn draw<D>(&self, ctx: &mut Context, param: D) -> GameResult
    where
        D: Into<DrawParam>,
    { ... }
}

All types that can be drawn on the screen implement the Drawable trait.

Required Methods

fn draw_primitive(
&self,
ctx: &mut Context,
param: PrimitiveDrawParam
) -> GameResult

Actually draws the object to the screen.

This is the most general version of the operation, which is all that is required for implementing this trait.

TODO: We could use a better name here.

fn set_blend_mode(&mut self, mode: Option<BlendMode>)

Sets the blend mode to be used when drawing this drawable. This overrides the general graphics::set_blend_mode(). If None is set, defers to the blend mode set by graphics::set_blend_mode().

fn get_blend_mode(&self) -> Option<BlendMode>

Gets the blend mode to be used when drawing this drawable.

Provided Methods

fn draw<D>(&self, ctx: &mut Context, param: D) -> GameResult where
D: Into<DrawParam>,

Draws the drawable onto the rendering target.

It just is a shortcut that calls draw_ex() with a default DrawParam except for the destination and rotation.

ctx - The Context this graphic will be rendered to.

dest - the position to draw the graphic expressed as a Point2.

rotation - orientation of the graphic in radians.

TODO: This should probably be removed, but can't actually be made generic on T where T: Into because we treat Drawable's as trait objects. ALSO TODO: Fix docs