Trait ggez::graphics::ShaderHandle

pub trait ShaderHandle<Spec: BackendSpec>: Debug {
    fn draw(
        &self, 
        _: &mut Encoder<Spec::Resources, Spec::CommandBuffer>, 
        _: &Slice<Spec::Resources>, 
        _: &Data<Spec::Resources>
    ) -> GameResult;

    fn set_blend_mode(&mut self, mode: BlendMode) -> GameResult;

    fn get_blend_mode(&self) -> BlendMode;
}

A trait that is used to create trait objects to abstract away the Structure type of the constant data for drawing

Required Methods

fn draw(
&self,
_: &mut Encoder<Spec::Resources, Spec::CommandBuffer>,
_: &Slice<Spec::Resources>,
_: &Data<Spec::Resources>
) -> GameResult

Draw with the current Shader

fn set_blend_mode(&mut self, mode: BlendMode) -> GameResult

Sets the shader program's blend mode

fn get_blend_mode(&self) -> BlendMode

Gets the shader program's current blend mode