Type Definition ggez::graphics::Canvas

type Canvas = CanvasGeneric<GlBackendSpec>;

A canvas that can be rendered to instead of the screen (sometimes referred to as "render target" or "render to texture"). Set the canvas with the ggez::graphics::set_canvas() function, and then anything you draw will be drawn to the canvas instead of the screen.

Resume drawing to the screen by calling ggez::graphics::set_canvas(None).

A Canvas allows creating render targets to be used instead of the screen. This allows graphics to be rendered to images off-screen in order to do things like saving to an image file or creating cool effects by using shaders that render to an image. If you just want to draw multiple things efficiently, look at SpriteBatch.

Methods

impl Canvas

pub fn new(
ctx: &mut Context,
width: u32,
height: u32,
samples: NumSamples
) -> GameResult<Canvas>

Create a new canvas with the given size and number of samples.

pub fn with_window_size(ctx: &mut Context) -> GameResult<Canvas>

Create a new canvas with the current window dimensions.

pub fn get_image(&self) -> &Image

Gets the backend Image that is being rendered to.

pub fn into_inner(self) -> Image

Destroys the Canvas and returns the Image it contains.