Type Definition ggez::graphics::Image

type Image = ImageGeneric<Resources>;

In-GPU-memory image data available to be drawn on the screen, using the OpenGL backend.

Under the hood this is just an Arc'ed texture handle and some metadata, so cloning it is fairly cheap; it doesn't make another copy of the underlying image data.

Methods

impl Image

pub fn new<P: AsRef<Path>>(context: &mut Context, path: P) -> GameResult<Image>

Load a new image from the file at the given path.

pub fn from_rgba8(
context: &mut Context,
width: u16,
height: u16,
rgba: &[u8]
) -> GameResult<Image>

Creates a new Image from the given buffer of u8 RGBA values.

pub fn to_rgba8(&self, ctx: &mut Context) -> GameResult<Vec<u8>>

Dumps the Image's data to a Vec of u8 RGBA values.

pub fn encode<P: AsRef<Path>>(
&self,
ctx: &mut Context,
format: ImageFormat,
path: P
) -> GameResult

Encode the Image to the given file format and write it out to the given path.

See the filesystem module docs for where exactly the file will end up.

pub fn solid(
context: &mut Context,
size: u16,
color: Color
) -> GameResult<Image>

A little helper function that creates a new Image that is just a solid square of the given size and color. Mainly useful for debugging.

pub fn width(&self) -> u32

Return the width of the image.

pub fn height(&self) -> u32

Return the height of the image.

pub fn get_filter(&self) -> FilterMode

Get the filter mode for the image.

pub fn set_filter(&mut self, mode: FilterMode)

Set the filter mode for the image.

pub fn get_dimensions(&self) -> Rect

Returns the dimensions of the image.

pub fn get_wrap(&self) -> (WrapMode, WrapMode)

Gets the Image's WrapMode along the X and Y axes.

pub fn set_wrap(&mut self, wrap_x: WrapMode, wrap_y: WrapMode)

Sets the Image's WrapMode along the X and Y axes.