Module ggez::mouse

Mouse utility functions.

Structs

MouseContext
Stores state information for the mouse, what little of it there is.

Enums

CursorState
Describes how winit handles the cursor.

MouseCursor
Describes the appearance of the mouse cursor.

Functions

get_cursor_state
Set whether or not the mouse is grabbed (confined to the window) or hidden (invisible).

get_cursor_type
Returns the current mouse cursor type of the window.

get_delta
Get the distance the cursor was moved during last frame, in pixels.

get_position
Get the current position of the mouse cursor, in pixels. Complement to set_position(). Uses strictly window-only coordinates.

set_cursor_state
Set whether or not the mouse is grabbed (confined to the window) or hidden (invisible).

set_cursor_type
Modifies the mouse cursor type of the window.

set_position
Set the current position of the mouse cursor, in pixels. Uses strictly window-only coordinates.