Struct ggez::Context

pub struct Context {
    pub conf: Conf,
    pub filesystem: Filesystem,
    pub continuing: bool,
    pub timer_context: TimeContext,
    pub audio_context: AudioContext,
    pub keyboard_context: KeyboardContext,
    pub mouse_context: MouseContext,
    pub gamepad_context: GamepadContext,
    pub default_font: Font,
    // some fields omitted
}

A Context is an object that holds on to global resources. It basically tracks hardware state such as the screen, audio system, timers, and so on. Generally this type is not thread- safe and only one Context can exist at a time. Trying to create another one will fail.

Most functions that interact with the hardware, for instance drawing things, playing sounds, or loading resources (which then need to be transformed into a format the hardware likes) will need to access the Context.

Fields
conf: Conf

The Conf object the Context was created with

filesystem: Filesystem

Filesystem state

continuing: bool

Controls whether or not the events loop should be running.

timer_context: TimeContext

Timer state

audio_context: AudioContext

Audio context

keyboard_context: KeyboardContext

Keyboard context

mouse_context: MouseContext

Mouse context

gamepad_context: GamepadContext

Gamepad context

default_font: Font

Default font

Methods

impl Context

pub fn load_from_conf(
game_id: &'static str,
author: &'static str,
default_config: Conf
) -> GameResult<(Context, EventsLoop)>

Tries to create a new Context by loading a config file from its default path, using the given Conf object as a default if none is found.

The game_id and author are game-specific strings that are used to locate the default storage locations for the platform it looks in, as documented in the filesystem module. You can also always debug-print the Context::filesystem field to see what paths it is searching.

pub fn quit(&mut self)

Terminates ggez::run() loop by setting Context::continuing to false.

pub fn process_event(&mut self, event: &Event)

Feeds an Event into the Context so it can update any internal state it needs to, such as detecting window resizes. If you are rolling your own event loop, you should call this on the events you receive before processing them yourself.